Introduce an Hebrew key map
Set the direction of the student's text according to the teacher, and not to the direction of the parent. Now Hebrew/Arabic/Farsi text will start at the right, and not 
left. However, the direction of the text is not set properly. To see the problem load ktouch using "ktouch --reverse" and when trying a lecture see where the space 
between words is (it's on the right for English texts. The exact problem happens when loading Hebrew in LTR UI. I am not sure how to fix this, since the text is printed 
by QGraphicsSimpleTextItem which does not provide a way to set the text "layout" direction. Next - look at the sources, maybe I can properly fix this.

BUG: 128268


svn path=/trunk/KDE/kdeedu/ktouch/; revision=748195