new training widget
Done in QML  and conceptually different than the old widget which will
it ultimately replace. It resembles an paper page on which the lecture
text ist displayed as an subtle placeholder. The text the user enters
replaces the placeholder, errors on his behalf are highlighted in red.
The page is embedded in a Flickable automically tracking the
current training line.

As for the keyboard widget new classes hold the training data.