SVG infrastructure and SVG graphics for training screen
We use Svg and SvgItem from org.kde.plasma.core to display SVG graphics.
Svg expects an absolute file path for non-themed (as our custom
graphics are) images, so we need a way to obtain these paths in QML.
This is achieved with the new context object for the QML view,
ContextView, which contains a method to lookup the path of a named
image.

As a test bed for the new infrastructe the background of the training
screen is now SVG themed.