configuration infrastructure
Consisting of three parts:
 * the actual core powered by KConfig XT
 * the configuration dialog
 * and the PreferencesProxy making the settings available in QML

Right now we have settings to configure wether the keyboard should be
visible during training, to set the tresholds for unlocking new levels
(not yet implemented) and the tinting colors for the keys indicating
the finger to use.

A few more settings will follow later, but not the plethora of settings
provided by KTouch 1.