ViewContext: create XEventNotifier without a parent widget
The XEventNotifier may not paint itself but it has non-zero dimensions
and consumes all mouse envents in its area. The XEventNotifier was created
as direct child of main window before obscurring the topleft corner of the
QML view. This prevented the profile switching button of the homescreen
to work poperly.