core: move calculation of lesson chars from DataIndex to Course
Each lessons specify which new characters it introduces, but for the
training what actually matters are all allowed characters, including
the characters from the preceding lessons.

Previously this was calculated at load time in DataAccess which was fine
fore readonly usage. Now we want to provide an editor for lessons, so
we update the characters for all affected lessons in the corresponding
course once neccessary.