trainer: reduce duration of switch screen animation
Atleast when switching to the home screen. The little pause before
training I actually consider to be a feature because it gives the
user time prepare mentalliy for the training. Also the longer animation
duration to the score screen is appropriate because this transition
is engaged implicitely after the lesson is completed. That way the user
has a real chance of actually figuring out what's happening.