i18n: don't translate lesson and course titles anymore
This was necessary for the old version of KTouch because the first
thing any user had to do before he could start training was to pick an
his keyboard layout and a matching course from the lists of all available
resources. For that it was important he was able to understand that
list, so they had to be localized.

Nowadays we detect the user's keyboard layout automatically and present
them only a choice for the courses matching this layout. We can assume
that he understands the language corresponding to the keyboard layout
his computer is set to. With that localization of keyboard layout and
course titles isn't necessary anymore, because the former are completly
invisible to the avarage user and the latter are very propably in a
language he understands.

This also elimates the long standing inconstancy that course titles
were localized but their contents not. Now everythins is presented in
the same language.

All course and keyboard layout for which the titles had native
translations available use these now.