core: store user resources in profiles database
Due to the non transactional nature of today's filesystems, it was
possible to get the data index and the actual user resource files out
of sync. If that happened the user had to rectify the situation himself
by editing XML files at a location he probobly didn't even know about.
This was clearly an unacceptable situation and that's fixed now.