drop hard dependency on X11
The keyboard layout detection depends on X11 and because of that
KTouch couldn't even build on other platforms. Now we provide a menu to
manually configure the keyboard layout on those platforms.

It's accassible via the main menu in the spot there on X11 the menu
entry to configure the system's keyboard layout would be. That one
would be defunct on other platforms anyways.