fix session managment behaviour
KTouch has two types of main windows: MainWindow (the trainer) and
ResourceEditor. The RESTORE macro the previously employed can't handle
multiple main window types. The the template function
kRestoreMainWindows can, but it ignores the special constraints between
MainWindow and ResourceEditor we have. So to fix that issue we duplicate
a little bit of code, but now session managment works as intended.