training screen: show hint when user is stuck
When the the user has entered a wrong character or is a the end of the
line and then did three key presses not changing the situation we show
to him the key has to press directly under the current line of the
lesson text.

This is especially useful when the keyboard layout visualization is
disabled in the settings.