resource editor: fix resizing of keys when zoom factor unequal one
When applying the scaled coordinates of the resize grips to the actual
key items the values were only implicitely rounded. This resulted in
an off-by-one error in the code responsible for retaining the dimensions
of the key not affected by current grip. These errors accumulated
quickly to rather big values because the code is executed for each
drag event. This is fixed now through explicit rounding.