buildsystem: use macro_display_feature_log for optional dependencies
KTouch has an optional dependendy on XKB for keyboard layout
auto-detection. I consider this feature to be of great use and
everyone should enable it if possible. The only reason the
dependency is optional is to enable compiling on platforms
where X11 isn't available. In the recent past I have observed
few cases where people compiled KTouch without XKB support even
on Linux. I hope this happens less often now that CMake gives a
report whether the support will be compiled in or not.