fix appearance of application with light-on-dark themes
In the most cases relying on theme colors instead of
fixed colors was enough, and in a few more cases for
elements with fixed background the opposite was right way
to achieve a good-looking result.

Most tricky proved to be the application background. I didn't
want to give up the Oxygen-provided background to ensure visual
consistency with the window decoration. Therefore I use Plasma's
background color but set the alpha accoardingly to the difference
between the two, if the colors are pretty similar the alpha is
zero.

BUG: 312082
FIXED-IN: 4.10 RC 2