resource editor: ensure stable sorting of items in the resource list
The items in the list are sorted lexically accordingly to their name.
This change affects the case if one had two items with the same name.
Naturally these two items where displayed together in the list, but
there was a really annoying usability bug: once you selected one item
the two item would swap their positions. That way the list looks
exactly the same as before suggesting the user that his selection
has been revoked. Which was not the case leading to unneccessary
confusion.

This fixed now by introducing a stable sort order even for items with
the same name.