core: store and load lessons IDs correctly from profile database
The serialisation/deserialisation code responsible for storing and
loading lessons from the database ommited the (very important) lesson
IDs. This fixed now and in the consequence statistics tracking
works now also for custom courses as intended.

The fix requires a change on the database schema so this commit
introduces a new schema version ('1.1') and the required migration
code.