training screen: better placement of key hint item
The hint key item is now displayed right below under the current
training line so the user can see it really well while also beeing able
to read the text he has to type. Additionally the item is now in front
of all over visual elements so the it gets never clipped by the keyboard
visualization.