home screen: extend course selector with a new page for custom lessons
The custom lessons selector is still just a placeholder but the course
selection logic works as intended. We treat the custom lessons as one
additional course available for each keyboard layout. As consequence,
there is now at least one course available for each keyboard layout, the
one with the custom lessons, so the fullscreen no-course warning which
we used previously has been retired.