store last selected lesson again
The function got lost when implementing the new custom lesson selector,
now it's there again, but realized a little bit different. This time we
store the lesson once a training session begins so we catch also the
case if the user selects the next lesson directly from the score
screen.