home screen: implement custom lesson managment
One can add, select, edit and delete custom lessons. For the delete
action an undo stack is available to get rid of the usual yes/no/cancel
question.

The only thing missing is that the lesson editor so far has no
information of the keyboard layout and theirfore can't decide which
characters are allowed.