custom lesson editor: implement allowed characters highlighting
We take the current layout as basis for this and collect the characters
of all the keys. This ensures the user has a chance to spot special
characters which might not be typable with his keyboard before he has
actually starting training.

In case KTouch has no data file for the user's current keyboard layout
all characters will be allowed.