home screen: fix bahavior when deleting last custom lesson
For some obscure reason the bindings chain stopped working properly
after deleting the last lesson. The result was that the just-deleted
lesson was still selected with the preview and the editing controls
visible.

We fix this by explicitly requesting an additional update of the
selected lesson.

This commit also includes some minor code clean-up creating during
debugging this issue.