Fix Keyboard Focus Loss Issue For Trainer
When the main window got reactivated by window manager interaction
the training area lost the keyboard focus. This makes it impossible
to resume training immediately, one has to click into the window one
more time to set the focus manually.

In fact the whole Qt Quick scene lost the focus. The scene in is
rendered in its own QWindow and embedded into the main window with
QWidget::createWindowContainer. Then we main window receives the focus
it doesn't delegate the focus automatically to the embedded QWindow.

Setting a focus policy of the container widget fixes this.

Also prevent tabbing out of the training area. This is not useful
thing to have.

BUG: 374701
FIXED-IN: 16.12.2