Fix Crash When Closing Resource Editor
We have to delete the QQuickView manually. The automatic deletion
is happening too late. At this stage the keyboard layout editor, which
we expose a context property, has no parent anymore. The QML deletes
relentlessly everything it sees without a parent. Which will inevitable
lead to double free and crash.