Home Screen: Uniform Spacing for Lesson Selector
It is not obvious how to achieve even spacing between and arround the
items of an grid view. Previously the spacing between two lesson items
was twice as large as the spacing between lesson items and the view
border. Now the spacing is uniform.

Also, the menu button above the lesson grid is now aligned to it.