Disable QML Precompilation by Default
The reduction of the time the app needs to launch because the
precompiled QML files is neglible, but we still get bug reports
because people try to run KTouch builds against newer Qt releases
which doesn't work. So from now on, this feauture must be turned
on explicitely to reduce further fallout.

CCBUG: 412677